Add two PHPUnits variant
