+++ composer.json
@@ -81,10 +81,10 @@
         "bower-asset/yii2-pjax": "~2.0.1"
     },
     "require-dev": {
-        "phpunit/phpunit": "4.8.34",
+        "phpunit/phpunit": "4.8.36|^7.0",
         "cebe/indent": "~1.0.2",
-        "friendsofphp/php-cs-fixer": "~2.2.3",
-        "johnkary/phpunit-speedtrap": "^1.0"
+        "friendsofphp/php-cs-fixer": "~2.2.3|~2.16.1",
+        "johnkary/phpunit-speedtrap": "^1.0|^3.0"
     },
     "repositories": [
         {
@@ -100,9 +100,6 @@
             "yii\\": "framework/",
             "yii\\cs\\": "cs/src/"
         }
-    },
-    "config": {
-        "platform": {"php": "5.4"}
     },
     "bin": [
         "framework/yii"
